+++ crates/rust-client/src/store/web_store/chain_data/js_bindings.rs
@@ -37,4 +37,10 @@ extern "C" {
 
     #[wasm_bindgen(js_name = insertChainMmrNodes)]
     pub fn idxdb_insert_chain_mmr_nodes(ids: Vec<String>, nodes: Vec<String>) -> js_sys::Promise;
+
+    // DELETES
+    // ================================================================================================
+
+    #[wasm_bindgen(js_name = pruneIrrelevantBlocks)]
+    pub fn idxdb_prune_irrelevant_blocks() -> js_sys::Promise;
 }