+++ crates/rust-client/src/store/web_store/js/chainData.js
@@ -1,4 +1,4 @@
-import { blockHeaders, chainMmrNodes } from "./schema.js";
+import { blockHeaders, chainMmrNodes, stateSync } from "./schema.js";
 
 // INSERT FUNCTIONS
 export async function insertBlockHeader(
@@ -173,6 +173,22 @@ export async function getChainMmrNodes(ids) {
     return results;
   } catch (err) {
     console.error("Failed to get chain mmr nodes: ", err.toString());
+    throw err;
+  }
+}
+
+export async function pruneIrrelevantBlocks() {
+  try {
+    const syncHeight = await stateSync.get(1).blockNum;
+    const allMatchingRecords = await blockHeaders
+      .where("hasClientNotes")
+      .equals("false")
+      .and((record) => record.blockNum !== 0 && record.blockNum !== syncHeight)
+      .toArray();
+
+    await blockHeaders.bulkDelete(allMatchingRecords.map((r) => r.blockNum));
+  } catch (err) {
+    console.error("Failed to prune irrelevant blocks: ", err.toString());
     throw err;
   }
 }