+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 * Updated Web Client README and Documentation (#808).
 * [BREAKING] Removed `script_roots` mod in favor of `WellKnownNote` (#834).
+* Updated the client so that only relevant block headers are stored (#828).
+* [BREAKING] Refactored the `StateSync` so that unverified notes are updated inside the component (#828).
 
 ## 0.8.1 (2025-03-28)